Add FPVue.key as default gs.key

+++ install.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
+set -e
+
 install_dir='/home/radxa/gs'
 script_dir=$(dirname $(readlink -f $0))
 cd $script_dir
@@ -12,7 +14,7 @@ fi
 [ -d ${install_dir}/udev-rules-bak ] || mkdir -p ${install_dir}/udev-rules-bak
 
 echo -e "\033[31mRemove /etc/NetworkManager/system-connections/*\033[0m"
-rm /etc/NetworkManager/system-connections/*
+[ $(ls -A "/etc/NetworkManager/system-connections/") ] && rm /etc/NetworkManager/system-connections/*
 echo -e "\033[31mDisable wifibroadcast.service wifibroadcast@gs.service\033[0m"
 systemctl disable wifibroadcast.service wifibroadcast@gs.service
 echo -e "\033[31mcopy gs.sh wfb.sh stream.sh to ${install_dir}/\033[0m"
@@ -29,6 +31,10 @@ cp 99-wfb.rules /etc/udev/rules.d/
 echo -e "\033[31msystemctl enable gs.service \033[0m"
 systemctl daemon-reload
 systemctl enable gs
+echo -e "\033[31mCopy FPVue.key to /config/gs.key and linked to /etc/gs.key\033[0m"
+cp FPVue.key /config/gs.key
+[ -f /etc/gs.key ] && rm /etc/gs.key
+ln -s /config/gs.key /etc/gs.key
 
 echo -e "\033[31mInstallation Complete, Configuration file is /config/gs.conf\033[0m"
 sync
